requirements.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
tensorflow
opencv-python